refactor(admin): add search by request queries for days
